+++ Giftbit.API/Model/Gift.cs
@@ -8,7 +8,7 @@
         public string ManagementDashboardLink { get; set; }
         public int RedeliveryCount { get; set; }
         public string CampaignId { get; set; }
-        public long PriceInCents { get; set; }
+        public int PriceInCents { get; set; }
         public string BrandCode { get; set; }
         public string CreatedDate { get; set; }
         public string DeliveryDate { get; set; }
